backend/requirements.txt
appdirs==1.4.3
asgiref==3.3.1
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.4.6
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
django-heroku==0.3.1
django-rest-knox==4.1.0
django-storages==1.11.1
djangorestframework==3.12.2
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
importlib-metadata==3.7.3
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
mysqlclient==2.0.3
packaging==20.3
pep517==0.8.2
progress==1.5
psycopg2==2.7.5
psycopg2-binary==2.8.6
pycparser==2.20
PyMySQL==1.0.2
pyparsing==2.4.6
pytoml==0.1.21
pytz==2021.1
requests==2.22.0
retrying==1.3.3
six==1.14.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.25.8
webencodings==0.5.1
whitenoise==5.2.0
zipp==3.4.1
